[Radoub] refactor: Promote ModelPreviewGLControl + helpers to Radoub.UI (#1908 prep, #2155)
Pure structural move per the cross-tool health plan
(NonPublic/Plans/2026-04-30-1996-1908-plan.md PR2). No rendering behavior
changes; QM creature preview continues to render identically.

Files moved:
- Radoub.UI/Controls: ModelPreviewGLControl, OpenGLShaderManager,
  MeshSkipHeuristic, SmoothGroupNormals, VertexWelder
- Radoub.UI/Services: ModelViewController, TextureService
- Radoub.UI.Tests: 4 helper test classes + TextureServiceHelperTests

csproj changes:
- Radoub.UI: AllowUnsafeBlocks=true, Silk.NET.OpenGL package reference
- Radoub.UI.Tests: <Using Include="Xunit" /> for global Xunit
- Radoub.Formats does NOT reference Radoub.UI (no circular dep)

Memory-leak hardening preserved verbatim:
- LRU caches (palette 128, rendered 64) in TextureService
- TRACE-level MDL parser logs (#2142)
- AppearancePanel.OnPanelUnloaded cleanup chain (#2034)
- ShaderManager.Cleanup() in OpenGlDeinit
- PreferBifTextures cache-clear path (#1867)

#2155 fix folded in: restored 12 themable
{DynamicResource FontSizeMedium|FontSizeSmall} bindings on
ItemDetailsPanel.axaml that were lost during PR #2151 extraction.

Tests: Radoub.UI 562/562 (was 508 + 54 moved), Quartermaster 1217/1217
(was 1271 - 54 moved), Fence 138/138. Solution build clean (0 errors).
