+++ Quartermaster/Quartermaster/Views/Panels/AppearancePanel.EventHandlers.cs
@@ -7,8 +7,8 @@ using Avalonia.Controls;
 using Avalonia.Input;
 using Avalonia.Interactivity;
 using Avalonia.Media;
-using Quartermaster.Controls;
 using Quartermaster.Services;
+using Radoub.UI.Controls;
 using Radoub.Formats.Logging;
 using Radoub.UI.Services;
 using Radoub.UI.Views;
