+++ Quartermaster/Quartermaster/Views/Panels/AppearancePanel.axaml
@@ -2,7 +2,7 @@
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
              xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
              xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-             xmlns:controls="using:Quartermaster.Controls"
+             xmlns:controls="using:Radoub.UI.Controls"
              mc:Ignorable="d" d:DesignWidth="900" d:DesignHeight="700"
              x:Class="Quartermaster.Views.Panels.AppearancePanel">
