+++ Quartermaster/Quartermaster/Views/Panels/AppearancePanel.axaml.cs
@@ -3,8 +3,8 @@ using System.Collections.Generic;
 using Avalonia.Controls;
 using Avalonia.Input;
 using Avalonia.Markup.Xaml;
-using Quartermaster.Controls;
 using Quartermaster.Services;
+using Radoub.UI.Controls;
 using Radoub.Formats.Logging;
 using Radoub.UI.Services;
 using Radoub.Formats.Mdl;
